Restored consistency in certain query responses
The WebSocket and RESTful API now returns the same JSON objects for

 * /find/metadata/by/id
 * /find/objects/by/id
 * /find/segments/by/id

Each of the corresponding RESTful action handler's now do have the same result
type as the corresponding WebSocket handlers have.

The main advantage is that both APIs do now have similar response
objects.
Further, there is a minor advantage in not sending top level arrays: A
few JSON processors cannot handle such documents and would therefore not
be able to be used in a cineast context.