Added new query type for lsc spatial query
The new type PARAMETERISED_LOCATION is more fexible than the existing
LOCATION type, as a) it allows to send ugly data-url-base64-json AND a
JSON-string (as a JSON value, doublequote escaping is a thing) and b) it
enables callers to pass the half similarity distance as an expression
(i.e. 1000/3). At this haversine-distance value, the similarity will be
0.5